Moved component to the regular components dir.
